+++ cakey-domain/src/main/java/com/cakey/cake/facade/CakeFacade.java
@@ -1,0 +1,13 @@
+package com.cakey.cake.facade;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class CakeFacade {
+
+    private final CakeRetriever cakeRetriever;
+
+    publi findMainImagesByStoreIds
+}
