+++ cakey-domain/src/main/java/com/cakey/cake/repository/CakeRepository.java
@@ -1,0 +1,7 @@
+package com.cakey.cake.repository;
+
+import com.cakey.cake.domain.Cake;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CakeRepository extends JpaRepository<Cake, Long> {
+}
