--- cakey-domain/src/main/java/com/cakey/cakeimage/dto/CakeImageDto.java
@@ -1,16 +1,0 @@
-package com.cakey.cakeimage.dto;
-
-import com.querydsl.core.annotations.QueryProjection;
-import lombok.Getter;
-
-@Getter
-public class CakeImageDto {
-    private final Long imageId;
-    private final String imageUrl;
-
-    @QueryProjection
-    public CakeImageDto(Long imageId, String imageUrl) {
-        this.imageId = imageId;
-        this.imageUrl = imageUrl;
-    }
-}
