--- cakey-domain/src/main/java/com/cakey/cakeimage/facade/CakeImageFacade.java
@@ -1,19 +1,0 @@
-package com.cakey.cakeimage.facade;
-
-import com.cakey.cakeimage.domain.CakeImages;
-import com.cakey.cakeimage.dto.CakeImageDto;
-import com.cakey.cakeimage.repository.CakeImagesRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-
-@Component
-@RequiredArgsConstructor
-public class CakeImageFacade {
-    private final CakeImagesRetriever cakeImagesRetriever;
-
-    public List<CakeImageDto> findMainImagesByStoreIds(final List<Long> storeIds) {
-        return cakeImagesRetriever.findMainImagesByStoreIds(storeIds);
-    }
-}
